feat: api error handling restruct
