+++ cognee/api/v1/exceptions/__init__.py
@@ -5,5 +5,9 @@ This module defines a set of exceptions for handling various data errors
 """
 
 from .exceptions import (
-    InvalidConfigAttributeError
+    InvalidConfigAttributeError,
+    DocumentNotFoundError,
+    DatasetNotFoundError,
+    DataNotFoundError,
+    DocumentSubgraphNotFoundError
 )
